+++ bootstrap3/templatetags/bootstrap3.py
@@ -11,7 +11,7 @@ from django.template.loader import get_template
 from ..bootstrap import (
     css_url, javascript_url, jquery_url, theme_url, get_bootstrap_setting
 )
-from ..html import render_link_tag
+from ..utils import render_link_tag
 from ..forms import (
     render_button, render_field, render_field_and_label, render_form,
     render_form_group, render_formset,
@@ -326,7 +326,7 @@ def bootstrap_field(*args, **kwargs):
 
     **example**::
 
-        {% bootstrap_form form_field %}
+        {% bootstrap_field form_field %}
     """
     return render_field(*args, **kwargs)
 
@@ -536,6 +536,7 @@ def bootstrap_pagination(page, **kwargs):
     **Parameters**:
 
         :page:
+        :parameter_name: Name of paging URL parameter (default: "page")
         :kwargs:
 
     **usage**::
@@ -553,7 +554,8 @@ def bootstrap_pagination(page, **kwargs):
 
 
 def get_pagination_context(page, pages_to_show=11,
-                           url=None, size=None, extra=None):
+                           url=None, size=None, extra=None,
+                           parameter_name='page'):
     """
     Generate Bootstrap pagination context from a page object
     """
@@ -600,8 +602,8 @@ def get_pagination_context(page, pages_to_show=11,
     if url:
         # Remove existing page GET parameters
         url = force_text(url)
-        url = re.sub(r'\?page\=[^\&]+', '?', url)
-        url = re.sub(r'\&page\=[^\&]+', '', url)
+        url = re.sub(r'\?{0}\=[^\&]+'.format(parameter_name), '?', url)
+        url = re.sub(r'\&{0}\=[^\&]+'.format(parameter_name), '', url)
         # Append proper separator
         if '?' in url:
             url += '&'
@@ -631,4 +633,5 @@ def get_pagination_context(page, pages_to_show=11,
         'pages_back': pages_back,
         'pages_forward': pages_forward,
         'pagination_css_classes': ' '.join(pagination_css_classes),
+        'parameter_name': parameter_name,
     }